separate port-list
