+++ catalystwan/endpoints/configuration/policy/list/port.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import PortList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.port import PortList, PortListEditPayload, PortListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class PortListEditPayload(PortList, PolicyListId):
-    pass
-
-
-class PortListInfo(PortList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyPortList(APIEndpoints, PolicyListEndpoints):
