+++ catalystwan/models/policy/list/port.py
@@ -1,0 +1,18 @@
+from typing import List, Literal
+
+from catalystwan.models.policy.lists import PolicyListBase
+from catalystwan.models.policy.lists_entries import PortListEntry
+from catalystwan.models.policy.policy_list import PolicyListId, PolicyListInfo
+
+
+class PortList(PolicyListBase):
+    type: Literal["port"] = "port"
+    entries: List[PortListEntry] = []
+
+
+class PortListEditPayload(PortList, PolicyListId):
+    pass
+
+
+class PortListInfo(PortList, PolicyListInfo):
+    pass
